Add Spotify fetching, refactor types
