+++ .env
@@ -1,0 +1,2 @@
+SPOTIFY_CLIENT_ID=
+SPOTIFY_SECRET=
