+++ package.json
@@ -15,11 +15,12 @@
 		"@catppuccin/palette": "^1.0.3",
 		"@fontsource-variable/inter": "^5.0.16",
 		"@fontsource-variable/jetbrains-mono": "^5.0.19",
+		"@spotify/web-api-ts-sdk": "^1.2.0",
 		"@sveltejs/adapter-auto": "^3.1.1",
 		"@sveltejs/adapter-static": "^3.0.1",
 		"@sveltejs/kit": "^2.5.0",
 		"@sveltejs/vite-plugin-svelte": "^3.0.2",
-		"@tabler/icons-svelte": "^2.46.0",
+		"@tabler/icons-svelte": "^2.47.0",
 		"@types/eslint": "8.56.2",
 		"@typescript-eslint/eslint-plugin": "^6.20.0",
 		"@typescript-eslint/parser": "^6.20.0",
@@ -36,7 +37,7 @@
 		"tslib": "^2.6.2",
 		"typescript": "^5.3.3",
 		"vite": "^5.0.12",
-		"wrangler": "^3.25.0"
+		"wrangler": "^3.26.0"
 	},
 	"type": "module"
 }
