+++ src/lib/data/discography/sources.ts
@@ -1,0 +1,6 @@
+export const sources = [
+    "bandcamp",
+    "spotify",
+    "appleMusic",
+    "youtube"
+] as const;
