+++ src/lib/types/discography/sources.ts
@@ -1,0 +1,9 @@
+import type { sources } from "$lib/data/discography/sources"
+
+export type Sources = {
+    [key in typeof sources[number]]?: string
+}
+
+export interface StreamSources extends Sources {
+    "stream": string
+}
